[JBPM-8380] Remove JBPMHelper
