+++ jbpm-test/src/main/java/org/jbpm/test/JbpmJUnitTestCase.java
@@ -16,7 +16,7 @@
 
 package org.jbpm.test;
 
-import static org.jbpm.test.JBPMHelper.txStateName;
+import static org.jbpm.test.JbpmJUnitBaseTestCase.txStateName;
 import static org.junit.Assert.*;
 
 import java.io.File;
