--- jbpm-test/src/main/resources/jBPM.properties
@@ -1,22 +1,0 @@
-# for creating a datasource
-persistence.datasource.name=jdbc/jbpm-ds
-persistence.datasource.user=sa
-persistence.datasource.password=
-persistence.datasource.url=jdbc:h2:tcp://localhost/~/jbpm-db
-persistence.datasource.driverClassName=org.h2.Driver
-
-# for configuring persistence of the session
-persistence.enabled=true
-persistence.persistenceunit.name=org.jbpm.persistence.jpa
-persistence.persistenceunit.dialect=org.hibernate.dialect.H2Dialect
-
-# for configuring the human task service
-taskservice.enabled=true
-taskservice.datasource.name=org.jbpm.services.task
-
-#defaults
-#taskservice.usergroupcallback=org.jbpm.services.task.identity.JBossUserGroupCallbackImpl
-#taskservice.usergroupmapping=classpath:/usergroups.properties
-
-#not used anymore
-#taskservice.transport=hornetq
